requirements.txt
pytesseract
Pillow 